Fix command param processing
QCommandLineParser requires the first element of the parameter list to
be the "name of the application", with equivalent for the command
approach this appliation uses being the command name.

Originally the command name was removed before returning the command's
parameters, but it must be kept as part of them to satisfy this
requirement.