Fix encoding at BPC > 1
With the current approach of chunking bits off of input data the last
chunk out of a given byte array may be incomplete if the chunk does not
align with a byte boundary of the original data. This is acceptible if
it happens with the very last byte woven into the medium as that will
just cause a slight amount of technically extra (but still within
expectations) distortion on the that final pixel when in Absolute mode
(Relative mode is entirely unaffected; however, if this occurs in the
middle of the full encode stream then PxWeaver will write an incomplete
chunk, effectively adding null padding to the image that is undetectable
by the decoder.

Therefore, all data must be encoded from one contiguous stream until a
better solution is devised.

See comment.