Move encode/decode into utility classes
Lots of changes but the scope is actually minimal. This simply wraps the
functionality of the encode function and EncodeSettings struct into one
class and the decode function into another. State is managed by those
classes via setters/getters instead of using a large number of arguments
for the encode/decode functions.