Completely overhaul implementation
Originally all IO was handled by the implementation of Encoder/Decoder,
but now IO is handled by modified versions of their innerworkings (i.e.
PxSequenceGenerator, ChSequenceGenerator, PxAccess, etc.), and:

Frame - Wraps the image in question, handles basic pixel traversal and
the reading/writing of meta pixel values.

Canvas - Created from a Frame. Augments Frame to include channel index
tracking and broader pixel access. Sub-classes QIODevice so that data
can arbitrarily be written to/read from the image at will without the
limitation of the original implementation in which all data had to
written/read in one shot.

Artwork - An encapsulation of the actual data to be used within an
image, as a base class, which allows for much easier implementation of
multiple types of PxCrypt images.

Decoder and Encoder are also now base classes to allow for multiple new
encoder/decoder types (tending to be paired with corresponding artwork
types).

There is still much to cleanup/optimize and it may be worthwhile to
combine Frame/Canvas together for simplicity.